Refactor Folder Structure of Ambassador Page
