+++ frontend/src/app/coworking/ambassador-home/ambassador-home.component.html
@@ -1,15 +1,1 @@
-<nav mat-tab-nav-bar [tabPanel]="tabPanel" color="accent">
-  <a
-    mat-tab-link
-    *ngFor="let link of links"
-    [routerLink]="link.path"
-    routerLinkActive
-    #rla="routerLinkActive"
-    [routerLinkActiveOptions]="{ exact: false }"
-    [active]="rla.isActive"
-    >{{ link.label }}</a
-  >
-</nav>
-<mat-tab-nav-panel #tabPanel>
-  <router-outlet></router-outlet>
-</mat-tab-nav-panel>
+<tab-container [links]="links" />
