+++ frontend/src/app/coworking/ambassador-home/ambassador-room/ambassador-room-list.component.html
@@ -1,10 +1,3 @@
-<banner-card
-  type="info"
-  [wide]="true"
-  icon="construction"
-  title="Page Under Construction"
-  body="This page's styling has not yet been updated to conform to Material 3." />
-
 <div *ngIf="upcomingReservations$ | async as reservations">
   <mat-pane
     class="content"
