--- frontend/src/app/coworking/ambassador-home/ambassador-room/ambassador-room.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AmbassadorRoomService } from './ambassador-room.service';
-
-describe('AmbassadorRoomService', () => {
-  let service: AmbassadorRoomService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AmbassadorRoomService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
