+++ frontend/src/app/coworking/ambassador-home/ambassador-xl/ambassador-xl-list.component.html
@@ -1,11 +1,4 @@
 <div>
-  <banner-card
-    type="info"
-    [wide]="true"
-    icon="construction"
-    title="Page Under Construction"
-    body="This page's styling has not yet been updated to conform to Material 3." />
-
   <mat-pane class="content" appearance="outlined">
     <mat-card-header>
       <mat-card-title>Reserve a Drop-in at the Welcome Desk</mat-card-title>
